Refactored types and added back Edition type
