+++ src/types/Cloudinary.ts
@@ -1,0 +1,3 @@
+export type CloudinaryPhoto = {
+  secure_url: string;
+};
