+++ src/types/Edition.ts
@@ -1,0 +1,9 @@
+import { Venue } from 'types/Venue';
+import { CloudinaryPhoto } from 'types/Cloudinary';
+
+export type Edition = {
+  isoString: string;
+  dateString: string;
+  venue: Venue;
+  photos?: CloudinaryPhoto[];
+};
