+++ src/types/Venue.ts
@@ -1,0 +1,11 @@
+export type VenueLocation = {
+  street1?: string;
+  street2?: string;
+  suburb?: string;
+  postcode?: string;
+};
+
+export type Venue = {
+  name: string;
+  location?: VenueLocation;
+};
